+++ script/DeployAll.s.sol
@@ -7,7 +7,6 @@ import "../src/DeployAll.sol";
 contract DeployAllScript is Script {
     function run() external {
         uint256 deployerPrivateKey = vm.envUint("PRIVATE_KEY");
-        vm.createWallet(deployerPrivateKey).addr;
         vm.startBroadcast(deployerPrivateKey);
         new DeployAll();
         vm.stopBroadcast();